pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "moltest"
version = "0.1.0"
description = "A CLI tool for running Molecule tests."
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"} 
authors = [
    {name = "Eddie Dunn", email = "gh@eddiedunn.com"} 
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Environment :: Console",
]
dependencies = [
    "click>=8.0",
    "packaging>=20.0",
    "colorama>=0.4",
    "PyYAML>=5.0"  # Added for YAML parsing
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-mock>=3.0.0",
]

[project.scripts]
moltest = "moltest.cli:cli"

[project.urls]
Homepage = "https://github.com/eddiedunn/moltest" 
Repository = "https://github.com/eddiedunn/moltest"

[tool.hatch.version]
path = "src/moltest/__init__.py"

[tool.hatch.build.targets.sdist]
exclude = [
  "/.github",
  "/docs",
  "/tests",
  "/.taskmaster",
  "/.vscode",
  "/.idea",
  "/*.log",
  "/*.cache",
  "/*.json", 
  "/*.md",   
  "/.env*",
  "Makefile",
  "*.sublime-project",
  "*.sublime-workspace",
]

[tool.hatch.build.targets.wheel]
packages = ["src/moltest"]
